cartpage updated

+++ database/migrations/2026_01_23_072431_create_products_table.php
@@ -22,13 +22,13 @@ return new class extends Migration
             $table->string('category');
             $table->string('subcategory');
             $table->integer('quantity')->default(1);
-            $table->bigInteger('regular_price');
-            $table->bigInteger('sale_price')->nullable();
+            $table->decimal('regular_price', 10, 2);
+            $table->decimal('sale_price', 10, 2)->nullable();
             $table->mediumText('description');
             $table->string('color')->nullable();
             $table->enum('product_type', ['top-selling', 'trending', 'featured', 'regular'])
                   ->default('regular');
-            $table->integer('item_weight');
+            $table->decimal('item_weight', 8, 2);
             $table->boolean('inStock')->default(true);
             $table->timestamps();
         });
